Remove store to /services/store
