+++ src/services/types/index.ts
@@ -1,3 +1,6 @@
+import { ThunkAction } from 'redux-thunk'; 
+import { Action, ActionCreator } from 'redux';
+import { store } from '../store'; 
 import { Dispatch } from 'redux';
 import { TConstructorState } from '../reducers/constructorItems';
 import { TErrorState } from '../reducers/error';
